+++ Sources/ArknightsClient/Features/Game/Domain/LaunchOptions.swift
@@ -17,31 +17,67 @@ enum GameDisplayMode: String, CaseIterable, Codable, Sendable {
 }
 
 enum GameResolution: String, CaseIterable, Codable, Sendable {
-	case sd = "640x480"
-	case hd = "1280x720"
-	case fullHD = "1920x1080"
-	case quadHD = "2560x1440"
 	case ultraHD = "3840x2160"
+	case quadHD = "2560x1440"
+	case cinemaFullHD = "2048x1080"
+	case wuxga = "1920x1200"
+	case fullHD = "1920x1080"
+	case wsxgaPlus = "1680x1050"
+	case uxga = "1600x1200"
+	case wsxga = "1600x1024"
+	case hdPlus = "1600x900"
+	case hdFourThree = "1440x1080"
+	case hdWide = "1360x768"
+	case sxgaPlus = "1280x960"
+	case wxga = "1280x800"
+	case wxgaWide = "1280x768"
+	case hd = "1280x720"
+	case scaledHD = "1176x664"
+	case xgaPlus = "1152x864"
+	case xga = "1024x768"
+	case svga = "800x600"
+	case ntscWide = "720x480"
+	case sd = "640x480"
 
 	var displayName: String { rawValue.replacing("x", with: " × ") }
 
 	var width: Int {
 		switch self {
-		case .sd: 640
-		case .hd: 1280
-		case .fullHD: 1920
-		case .quadHD: 2560
 		case .ultraHD: 3840
+		case .quadHD: 2560
+		case .cinemaFullHD: 2048
+		case .wuxga, .fullHD: 1920
+		case .wsxgaPlus: 1680
+		case .uxga, .wsxga, .hdPlus: 1600
+		case .hdFourThree: 1440
+		case .hdWide: 1360
+		case .sxgaPlus, .wxga, .wxgaWide, .hd: 1280
+		case .scaledHD: 1176
+		case .xgaPlus: 1152
+		case .xga: 1024
+		case .svga: 800
+		case .ntscWide: 720
+		case .sd: 640
 		}
 	}
 
 	var height: Int {
 		switch self {
-		case .sd: 480
-		case .hd: 720
-		case .fullHD: 1080
-		case .quadHD: 1440
 		case .ultraHD: 2160
+		case .quadHD: 1440
+		case .cinemaFullHD, .fullHD, .hdFourThree: 1080
+		case .wuxga, .uxga: 1200
+		case .wsxgaPlus: 1050
+		case .wsxga: 1024
+		case .hdPlus: 900
+		case .hdWide, .wxgaWide, .xga: 768
+		case .sxgaPlus: 960
+		case .wxga: 800
+		case .hd: 720
+		case .scaledHD: 664
+		case .xgaPlus: 864
+		case .svga: 600
+		case .ntscWide, .sd: 480
 		}
 	}
 }
@@ -53,6 +89,7 @@ struct GameLaunchOptions: Codable, Sendable, Equatable {
 	var usesHighResolutionMode: Bool = true
 	var usesMetalPerformanceHUD: Bool = false
 	var usesGameMode: Bool = false
+	var synchronizationMode: WineSynchronizationMode = .msync
 
 	static let `default` = GameLaunchOptions(
 		displayMode: .windowed,
@@ -60,7 +97,8 @@ struct GameLaunchOptions: Codable, Sendable, Equatable {
 		usesGameSettings: true,
 		usesHighResolutionMode: true,
 		usesMetalPerformanceHUD: false,
-		usesGameMode: false
+		usesGameMode: false,
+		synchronizationMode: .msync
 	)
 
 	private enum CodingKeys: String, CodingKey {
@@ -70,6 +108,7 @@ struct GameLaunchOptions: Codable, Sendable, Equatable {
 		case usesHighResolutionMode
 		case usesMetalPerformanceHUD
 		case usesGameMode
+		case synchronizationMode
 	}
 
 	init(
@@ -78,7 +117,8 @@ struct GameLaunchOptions: Codable, Sendable, Equatable {
 		usesGameSettings: Bool = true,
 		usesHighResolutionMode: Bool = true,
 		usesMetalPerformanceHUD: Bool = false,
-		usesGameMode: Bool = false
+		usesGameMode: Bool = false,
+		synchronizationMode: WineSynchronizationMode = .msync
 	) {
 		self.displayMode = displayMode
 		self.resolution = resolution
@@ -86,6 +126,7 @@ struct GameLaunchOptions: Codable, Sendable, Equatable {
 		self.usesHighResolutionMode = usesHighResolutionMode
 		self.usesMetalPerformanceHUD = usesMetalPerformanceHUD
 		self.usesGameMode = usesGameMode
+		self.synchronizationMode = synchronizationMode
 	}
 
 	init(from decoder: any Decoder) throws {
@@ -112,6 +153,11 @@ struct GameLaunchOptions: Codable, Sendable, Equatable {
 				Bool.self,
 				forKey: .usesGameMode
 			) ?? false
+		synchronizationMode =
+			try container.decodeIfPresent(
+				WineSynchronizationMode.self,
+				forKey: .synchronizationMode
+			) ?? .msync
 	}
 
 	/// Unity standalone-player arguments supported by the Windows client.